unlink file

+++ app/Http/Controllers/WelcomeVideoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\WelcomeVideo;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
 class WelcomeVideoController extends Controller
@@ -117,15 +118,17 @@ class WelcomeVideoController extends Controller
             if (!$video) {
                 return response()->json(["error" => "Video not found"], 404);
             }
+
             $filePath = $video->file;
-            // return response()->json(["file" => $filePath]);
-            $video->delete();
-            // Use unlink for direct file deletion
-            // if (file_exists($filePath)) {
-            //     unlink($filePath);
-            // }
-            Storage::delete($filePath);
-            return response()->json(["message" => "Welcome video successfully deleted"], 200);
+
+            // Use File::delete to delete the file
+            if (File::delete($filePath)) {
+                // If the file is deleted successfully, you can proceed to delete the database record
+                $video->delete();
+                return response()->json(["message" => "Welcome video successfully deleted"], 200);
+            } else {
+                return response()->json(["error" => "Error deleting file from storage"], 500);
+            }
         } catch (\Throwable $th) {
             return response()->json(["error" => $th->getMessage()], 400);
         }
